fix navigation, split sites
