+++ frontend/src/routes/about/spenden/+page.svelte
@@ -1,13 +1,52 @@
 <script lang="ts">
-	import header from '$lib/assets/heroshots/karlsruhe_luft.jpeg';
-	import Donate from '$lib/components/Donate.svelte';
-	import HeroShot from '$lib/components/HeroShot.svelte';
 </script>
-
-<HeroShot imgSrc={header} />
 
 <main class="container mx-auto">
 	<div class="flex flex-col gap-8 px-4 py-20 xl:px-80">
-		<Donate />
+		<h1 class="text-4xl font-bold uppercase md:text-5xl">Unterstützung</h1>
+
+		<div class="grid gap-8 md:grid-cols-[1fr_3fr]">
+			<h2 class="text-3xl font-bold uppercase">Gebet</h2>
+			<p>
+				Bete, dass Gottes Liebe an unserer Hochschule sichtbar wird, Gott uns als Hochschulgruppe
+				gebraucht und Studierende Jesus kennen lernen können.
+			</p>
+
+			<h2 class="text-3xl font-bold uppercase">Mitarbeit</h2>
+			<p>
+				Du hast Interresse bei uns in der Gruppe mitzuarbeiten? Du bist herzlich willkommen!
+				Kontaktiere uns, schau bei einem unserer Gruppenabende vorbei und steig mit ein. Wir freuen
+				uns auf dich.
+			</p>
+
+			<h2 class="text-3xl font-bold uppercase">Spenden</h2>
+
+			<div class="grid gap-8">
+				<p>
+					Neben Engagement und Gebet lebt unsere Arbeit auch von finanzieller Unterstützung. Wir
+					freuen uns, wenn du uns unterstützt. Weitere Infos zur Finanzierung der SMD findest du
+					hier.
+				</p>
+
+				<ul class="font-bold">
+					<li>SMD e. V.</li>
+					<li>Evangelische Bank, Kassel</li>
+					<li>IBAN: DE75 5206 0410 0000 8004 57</li>
+					<li>Verwendungzweck: H-SMD Gruppe Karlsruhe</li>
+				</ul>
+
+				<p class="font-bold">
+					Wir danken dir für deine Spende, sie hilft uns dabei auch in Zukunft diese wichtige Arbeit
+					fortzuführen.
+				</p>
+			</div>
+		</div>
+		<div>
+			Wenn du über die aktuelle Aktionen und Anliegen der Gruppe informiert werden möchtest, kannst
+			du uns gerne unter dieser Email Adresse <a href="mailto:leiter@smd-karlsruhe.de"
+				>leiter@smd-karlsruhe.de</a
+			> nach unserem Freundesbrief nachfragen. Im Freundesbrief stehen alle aktuellen und wichtigen Informationen
+			sowie unsere Gebetsanliegen der Gruppe.
+		</div>
 	</div>
 </main>
