+++ frontend/src/routes/new/erstsemester/+page.ts
