+++ frontend/src/routes/new/wohnen/+page.svelte
@@ -1,0 +1,25 @@
+<script lang="ts">
+	import kaheim from '$lib/assets/logos/kaheim.png';
+</script>
+
+<main class="main">
+	<section class="pad">
+		<h2 class="pt-8">Wohnen</h2>
+		<div class="flex">
+			<p>
+				<a class="float-right align-middle" href="https://kaheim.de">
+					<img src={kaheim} class="h-12 self-center" alt="kaheim logo" />
+				</a>
+				Auf der Suche nach einer Wohnung? Dann schau doch mal bei
+				<a href="https://kaheim.de">kaheim.de</a> vorbei! Hier findest du eine Zimmerbörse für
+				christliche WGs in Karlsruhe. Wenn du ein Zimmer suchst oder in deiner WG Zimmer frei hast
+				und Mitbewohner suchst, kannst die Einträge dort durchschauen und/oder selbst einen
+				erstellen. Wir betreiben die Plattform selbst in Koorperation mit dem SfC (Studierende für
+				Christus), daher findest du alle Zimmerangebote, von denen wir wissen, auf dieser Seite.
+				<br />
+				Darüber hinaus kannst du dich auch mal beim
+				<a href="https://www.sw-ka.de/de/wohnen/"> Studierendenwerk</a> umschauen.
+			</p>
+		</div>
+	</section>
+</main>
